Reduce consistency requirement for token lookup.
Reduce the consistency requirement from Quorum to
One for push token lookups in the main async code
path for sending (native) push notifications to
increase chances of delivery in the presence of
network problems (and possibly slightly reduced
latency).

Note: Other code paths for push token lookups
still need to use Quorum consistency as the
lookups happen in the context of possible
token updates or removal, i.e. a (strongly)
consistent read is typically necessary for
correct operation and / or to fulfill the
API contract.