New field in NewUser for storing a UUID pre-fabricated.
In spar, we will need to be able to pick a UUID for a user we want to create.
Otherwise we would not be able to find dangling users after crashed creation
attempts.

Adding an optional field for this to the `NewUser` type is a small, reasonably
safe change: the two place where they are pulled from the network are here:

```
$ git grep -Hn NewUser
[...]
services/brig/src/Brig/API.hs:1150:    NewUserNoSSO new <- parseJsonBody req
services/brig/src/Brig/API.hs:1189:    (uData :: NewUser)  <- parseJsonBody req
[...]
```

We only allow the 'UUID' to be chosen by the creator in the 'NewUser' case;
the 'NewUserNoSSO' case fails to parse if anybody tries.