map x1 to defRetry and x5 to eagerRetry
eagerRetryHandlers in all the places where x5 is currently used seems safe,
since lightweight transactions, counters and operations on lists are not
in use in any of the x5 retry statements. The only query which 'looks'
unsafe is
https://github.com/wireapp/wire-server/blob/develop/services/brig/src/Brig/Unique.hs#L56,
however it operates on sets, not on lists; so it is safe to use
eagerRetry here.