Actually remove bad test.
I think this is not worth saving.  the 'bulkPush' test above the one
I delete here covers strictly more code than this test, and it works
both in the k8s setting and when running `/services/integration.sh`
(see haddocks for 'bulkPush').  Saving the test would require to read
the presences and fan out the bulk push request over the resources
(think: cannon IP addresses) on which the channels ended up being
connected.  It's not hard, but the test code would be harder to read
and maintain in the future, for very little gain.