Tweak test cases.
This is changing behavior that can be observed by the client, but I
think we should do this anyway.  The only way to avoid it would be to
keep the clients table around, but that is solely for the crypto keys,
which are not used any more.  So we would keep production code merely
because a test case sais it exists.