Remove dead code.
The exception handlers in the warp settings could never be triggered
since `catchErrors` would transform all exceptions into HTTP responses
before the `Application` got passed to warp.

Alternatively, we could drop `catchErrors`, but (1) with that we are
more flexible to process return values as well if we need to, and (2)
dropping the currently unused handlers from the warp settings is the
smaller change.