Fix compilation errors.
Until now, we have built against #6, but the release is based on
the alternative in #7.  tinylog does not split up `new` into a
version with and a version without effects, but reads the
`Settings` for a flag that determines whether effects should be
had or not.  #7 also lacks some helpers that were not general
enough, so I've copied them verbatim from #6 into this commit.