add legalhold disabled event (#794)
Adds a new event that gets sent to all members of a team when a legalhold device is disabled by an admin for a specific user

Before, devices did not know when legalhold would go from pending to disabled. Even when legalhold was disabled before the user approved it, the notification to approve would still show up. We now make sure that we send a notification when legalhold gets disabled, such that the client can react appropriately.