Fix: disambuate rest api.
The existing routes could only be distinguished by their query
arguments, which appearently is not enough for servant-swagger (and
may not be enough for whatever standard covers that question).  adding
constant path segments makes things more robust.

(I think the reason why wai-routes didn't have this issue is that
there is a combinator that allows the user to put the
parameter *either* in the query *or* in the path.  in the path,
disambiguation seems to work based on types, even though it's less
straight-forward and seems less robust to me.)