Make these values part of the "Settings" object
Problem now is that they're mandatory. We should have a better
story for introducing new config options. This is not very ergnomic. I'm
not sure 'where' in the codebase to set the default. I could make a
newtype wrapper with a custom FromJSON instance but this is conflating
encoding with business logic, which I do not like.