stack:  Split nix stuff into build deps and runtime deps.
Stack started not liking being called from a nix-shell anymore, because
stack wants to create the nix-shell itself so it does some
auto-detection and crashes.  To trick it into doing the right thing
again, we only use the nix-shell for non-build-deps and just for dev
tools, then stack itself will call another nix-shell file that just
handles build deps for the stack project.