Do not send events to team members to be deleted
Events about deleted conversations do not need to be sent to everyone
when a team is deleted. Indeed, it is harmful in that it leads to performance problems:
For large teams with many conversations, notifying
every team member about that fact that every conversation has been
deleted creates a lot of events.
None of these events are useful, since on a team deletion, the accounts
of every user are deleted, i.e. in practice these events cannot be read.

This PR fixes the bug that was introduced in #849