Don't provision SAML user twice. Also document carefully
Because it is hard to write tests for this piece of code, as we can't
simulate failures or mock out components in our test suite, I instead
write very careful comments and hope a reviewer will catch any logical
inconsistencies.

If anybody can come up with a good way to test this, I'd love to hear
it, as just trying this out 'live' is a bit scary.