Implement list users with efficient filter
Two bugs remain in the implementation (and hence 3 failing test cases):

1. SCIM should not return 404, but empty list if filter doesn't match
    MaybeT transformer can be used to write this elegantly I think. This
    should take a morning to rewrite.
2. 'testFindNonProvisionedUser' fails because createTeamMember doesnt
  assign a handle to an SSO user. Tiago thinks this should be a case
  we should be aware of.  I thought it was safe to assume that
  all SSO users have handles. So maybe we should log a warning when
  a user doesnt have a handle; but not error out
  we should also add a function that creates a team member with a
  handle. not too hard. so that the test actually works again

Not a bug, but a test that can not be written currently:

3. search by 'externalId' on non-provisioned users cannot be tested as
createTeamMember creates an "inconsisent" user. E.g. a brig user that
claims it's an SSO user; but is not part of the sso table on the spar
side. I should ask @fisx tomorrow how to make it stop doing that. As I
assume this is also important for other Spar tests? There is some hints
in the docs on how to fix it (e.g. by adding an internal endpoint to
spar).