Remove feature to opt out of searchablity.
The clients never supported this, and the flag meant more complications in
implementing search for tier-1 teams. So, for now, this feature is removed.

This commit does not include change in DB schema as we want to make sure we keep
the DB schema backwards compatible for at least a release in order to make sure
any release can be rollbacked.