Add gundeck
Yes ./dist is problematic.  Cabal the library tries to write to it,
but in our case it's a symlink, which points to the nothingness of space
(because the build is chrooted).

./dist isn't used for cabal stuff in our project, but for deb files and
such.  We shouldn't have named it ./dist as this confuses cabal.

For now i'm deleting this files, but I should be friendly to collegues
and not break their old workflows, so in the long term, I should
fix our Makefiles to not use ./dist but ./dist-wire or something related

The reason that this symlink exists in the first place is due to a bug
in our Makefiles that was worked around by placing the symlink...