Remove feature flags and CPP from core libs.
This is a noisy commit since I ran ormolu right away, but you only
need to know that the packages have more dependencies now even if the
flags are turned off, and that CPP is not enabled any more in all
modules that have been changed.  this is the only change.

This is a good change independently of ormolu, because turning off the
flags wasn't tested anywhere and didn't quite work anyway.