Compile nginx with libzauth  using nix (#988)
* Compile nginx with libzauth using nix/niv

Noteworthy:

both vts and moreheaders are packaged in nixpkgs, so no need
to fetch those git submodules... but perhaps there are version
conflicts. we should check! :)

I also use niv for version management of nixpkgs. instead of hacky own
thing.

* Git ignore rust build artifacts

* Actually add rust build artifacts to gitignore.

It's in the top-level gitignore of wire-server, but not in the one in
libzauth

This is all a bit confusing. We should unify everything in one top-level
gitignore. That makes things a lot less complicated. However,
this is an impurity to tackle in another commit