Fix .gitignore shenanigans in Nix (#1002)
Before, for building libzauth we were only considering the .gitignore
in the ./lib/libzauth directory, (ignoring the toplevel one) which
didn't contain a Cargo.lock ignore (and shouldn't! lockfiles should be
comitted so builds are reproducible!).  Then we deleted the local
gitignore
which tripped up the nix build as the gitignore didn't exist anymore.
We now use the toplevel .gitignore, but remove the global Cargo.lock
ignore as Cargo.lock's are important to be comitted (and were already!).
For some reason we do not have a Cargo.lock for ./libs/libzauth/libzauth
but only ./libs/libzauth/libzauth-c so I kept it like that.  libzauth-c
is the thing we directly build, and then pulls in libzauth as a
dependency, so  locking at just libzauth-c is sufficient and is also the
lockfile we had in source control already anyway.