stack snapshot 3.0.
This does not change any dependencies from 2.2.

I broke 2.* a while ago by editing the files, and thus learned
that even editing typos in comments is a bad idea.  the easiest
way out is to abandon the old snapshot and clone it into a new
one with the same content.