Move away from snapshots completely
We fold everything into stack.yaml

I really do not see any difference anymore, and with pantry this change
should be completely safe to make. We can now freely change stack.yaml
and get the same caching guarantees as before. (That is; if we have a
cache from a previous build that share 95% of the dependencies with the
new stack.yaml, then CI will just gladly pick this up due to how Pantry
works)

Again as noted before, what to do with the old 'snapshots/' folder I do
not know yet. We need to keep it around to keep old builds working
as they refer to it not by commit id but by 'develop' branch which is a
moving target.