Ignore bounds for some dependencies
We were already doing this implicitly because we used custom snapshot
before. Now we have to do it explicitly.

For good records, I attached the bounds issues to this commit.
Note that at least multihash is maintained by ourselves. we should thus
be able to solve the bounds issues for it

Error: While constructing the build plan, the following exceptions were encountered:

In the dependencies for HaskellNet-0.5.1:
    mime-mail-0.5.0 from stack configuration does not match >=0.4.7 && <0.5  (latest matching version is 0.4.14)
needed due to brig-1.35.0 -> HaskellNet-0.5.1

In the dependencies for HaskellNet-SSL-0.3.4.1:
    connection-0.3.1 from stack configuration does not match >=0.2.7 && <0.3  (latest matching version is 0.2.8)
needed due to brig-1.35.0 -> HaskellNet-SSL-0.3.4.1

In the dependencies for bloodhound-0.16.0.0:
    containers-0.6.0.1 from stack configuration does not match >=0.5.0.0 && <0.6  (latest matching version is 0.5.11.0)
    http-client-0.6.4 from stack configuration does not match >=0.4.30 && <0.6  (latest matching version is 0.5.14)
needed due to brig-1.35.0 -> bloodhound-0.16.0.0

In the dependencies for multihash-0.1.6:
    base-4.12.0.0 from stack configuration does not match >=4.7 && <4.12  (latest matching version is 4.11.1.0)
needed due to brig-1.35.0 -> multihash-0.1.6

In the dependencies for stm-hamt-1.2.0.4:
    primitive-0.6.4.0 from stack configuration does not match >=0.7 && <0.8  (latest matching version is 0.7.0.1)
    primitive-extras-0.7.1.1 from stack configuration does not match >=0.8 && <0.9  (latest matching version is 0.8)
needed due to spar-0.1 -> stm-hamt-1.2.0.4