Reproducibility (#1027)
* Add stack.yaml.lock and commit cabal files

By comitting stack.yaml.lock our build plan is fully reproducible. We've
had several cases where people could not reproduce a build on develop
before either due to cabal revisions or due to people accidentally
touching snapshot files.  stack.yaml.lock is a fixed build plan with
hashes derived from your snapshots and stack.yaml

See
https://github.com/commercialhaskell/stack/blob/master/doc/lock_files.md
for more info

We also commit the cabal files. Not committing them also reduces
reproducibility and the Stack people admit this now as well
https://tech.fpcomplete.com/blog/storing-generated-cabal-files In the
next stack release generating them will be the default.

* Create a new snapshot that pins correct version of hsaml2

Also pin correct version of saml2-web-sso

I took the liberty to move away from our system of having snapshots in
 ./snapshots/*.yaml and have new snapshots refer to previous ones. This
 was a workaround for getting better caching in Stack 1.0 which is
 simply not needed anymore.  We have a new single `wire-snapshot.yaml`
 file which you are always free to modify. Because Stack 2.0 uses
 content-addressed storage, making changes to this file will still cause
 most packages to be re-used.

With  Stack 2.0 and the introduction of pantry, there is no semantic
difference anymore between `extra-deps` in your `stack.yaml` and custom
snapshots. We could thus fold everything into a single `stack.yaml` and
keep the same benefits.

However custom snapshot format format gives you a bit more control (you
can for example also remove packages from a snapshot, which is not
possible in `stack.yaml`)
[Source](https://docs.haskellstack.org/en/stable/pantry/#snapshots)

So we can just have one single `wire-snapshot.yaml`
where we define external deps (and fold all the changes of all our
snapshots into that single file). We do not have to make
distinctions anymore between  "Things that change often go into
`stack.yaml` and things that change seldom go into `snapshot.yaml`" as
under the hood they're all the same build and caching mechanism now.

We also do not have to worry about not editing this file anymore. They
don't have to be treated as immutable as before. So no need to create a
`blah-2.0.yaml` referring to `blah-1.0.yaml` to work around the fact
that stack doesn't refetch snapshots after edited.  From Stack
documentation

>  They are assumed to be mutable, so you are free to modify it. We
> detect that the snapshot has changed by hashing the contents of the
> involved files, and using it to identify the snapshot internally. It is
> often reasonably efficient to modify a custom snapshot, due to stack
> sharing snapshot packages whenever possible.

We keep the ./snapshots folder for now; as our snapshots do not refer to
themselves by relative path but by absolute URL which mentions the
branch name and github repository.  This is a bit of a shame as this
means we can never remove them as this will break older builds :(

I do not know how much we care that old builds keep working, but having
them work by promising some url to a mutable artifact stays around was
never very good for reproducibility so maybe we should not care..
However; lets break this in a different PR after heavy discussion :)

* Move away from snapshots completely

We fold everything into stack.yaml

I really do not see any difference anymore, and with pantry this change
should be completely safe to make. We can now freely change stack.yaml
and get the same caching guarantees as before. (That is; if we have a
cache from a previous build that share 95% of the dependencies with the
new stack.yaml, then CI will just gladly pick this up due to how Pantry
works)

Again as noted before, what to do with the old 'snapshots/' folder I do
not know yet. We need to keep it around to keep old builds working
as they refer to it not by commit id but by 'develop' branch which is a
moving target.

* Ignore bounds for some dependencies

We were already doing this implicitly because we used custom snapshot
before. Now we have to do it explicitly.

For good records, I attached the bounds issues to this commit.
Note that at least multihash is maintained by ourselves. we should thus
be able to solve the bounds issues for it

Error: While constructing the build plan, the following exceptions were encountered:

In the dependencies for HaskellNet-0.5.1:
    mime-mail-0.5.0 from stack configuration does not match >=0.4.7 && <0.5  (latest matching version is 0.4.14)
needed due to brig-1.35.0 -> HaskellNet-0.5.1

In the dependencies for HaskellNet-SSL-0.3.4.1:
    connection-0.3.1 from stack configuration does not match >=0.2.7 && <0.3  (latest matching version is 0.2.8)
needed due to brig-1.35.0 -> HaskellNet-SSL-0.3.4.1

In the dependencies for bloodhound-0.16.0.0:
    containers-0.6.0.1 from stack configuration does not match >=0.5.0.0 && <0.6  (latest matching version is 0.5.11.0)
    http-client-0.6.4 from stack configuration does not match >=0.4.30 && <0.6  (latest matching version is 0.5.14)
needed due to brig-1.35.0 -> bloodhound-0.16.0.0

In the dependencies for multihash-0.1.6:
    base-4.12.0.0 from stack configuration does not match >=4.7 && <4.12  (latest matching version is 4.11.1.0)
needed due to brig-1.35.0 -> multihash-0.1.6

In the dependencies for stm-hamt-1.2.0.4:
    primitive-0.6.4.0 from stack configuration does not match >=0.7 && <0.8  (latest matching version is 0.7.0.1)
    primitive-extras-0.7.1.1 from stack configuration does not match >=0.8 && <0.9  (latest matching version is 0.8)
needed due to spar-0.1 -> stm-hamt-1.2.0.4