Tune tokenization for better results
While indexing tokenize more things, like letters and symbols and while
searching only search for things broken by whitespaces. This still leaves a
little gap of the characters which are not letters, digits, symbols,
punctuations or whitespace. Unfortunately such characters exist all I have found
were non-digit numbers like "U+2182 Roman Numeral Ten Thousand", the ICU
transliteration fails to transliterate it and it is none of the above
categories. This means users with these characters might be harder to find, they
will still be searchable by exact name/handle, so maybe this is not a big
problem in real life. However, in tests we will have to figure out a way to not
run into this problem and still be able to test with different names.