Use more aligned analyzers for indexing and searching
Earlier they were not aligned and caused search to produce tokens which were not
in index. From now this should not happen. Search will still produce one token
for each word in the query, but that word will definitely be in the index (of
course only if the name was index)