Implement better prefix search for name/handle (#1052)
Prefix search was not exhaustive. This change makes it exhaustive using the `edge_ngram` filter. For this, a new index is required. Search will continue to work with old index as well, it will still not be exhaustive. To use the new search a new index must be created and data has to be migrated across.

This commit also includes these changes:
* Add more tests to cover searching
* Add docs of how the search works
* Add docs on how to migrate to a new index
* Add command to `brig-index` binary to help with migation