Simplify team feature database interface.
There was one Data.* module per column in the team_features table.
That may make sense at some point in the distant future, but as long
as all those columns are just booleans, having one haskell module to
access any of them in a uniform way saves a lot of boilerplate.