Make cargohold use amazonka instead of aws (#1157)
* squash https://github.com/wireapp/wire-server/pull/483

* remove all usages of aws

* support s3DownloadEndpoint config

* stack.yaml: pin updated fork of amazonka

The update to stack.yaml just takes all amazonka packages from the pinned fork.
They depend on core, so need to be rebuilt anyways.
We also drop elb, redshift, route53

* depend on OpenSSL again

This is mostly done for consistency with the other services. If we want to change it,
we should change it everywhere.

* remove unnecessary dependencies on binary and blaze-builder

* Add comment on how to use cloudfront

Co-authored-by: Tiago Loureiro <tiago.loureiro@gmail.com>