Add script to rebase branch onto formatting changes (#1180)
Upgrading Ormolu can be painful, because the style will often change in
a lot of places (see for example #1145). Any active branches/PRs will
then have to rebase onto these changes and resolve the resulting
conflicts.
It becomes a bit less cumbersome after first squashing all commits of
the branch into a single one, but that's often not desirable.

This script will automate the rebasing process, keeping the (linear)
history of intact and making the commits appear as if the changes had
been applied onto the newly-formatted version all along.