This is probably bad.
another dead end: if i update the handle in an invitation, i need to
deal with a number of inconsistencies and race conditions in cassandra;
if i don't, what should i do when updating a user?  just return the
old handle?  and in case there wasn't one and we now construct a default,
what do i do there?

and is it even a good idea to update the default handle in the function
that looks like it's just doing a lookup?