Allow SCIM without saml (#1200)
documentation

Before this PR, scim tokens could only be added to teams that already had exactly one SAML IdP. Now, we also allow SAML-less teams to have SCIM provisioning. This is an alternative to onboarding via team-settings and produces user accounts that are authenticated with email and password. (Phone may or may not work, but is not officially supported.)

The way this works is different from team-settings: we don't send invites, but we create active users immediately the moment the SCIM user post is processed. The new thing is that the created user has neither email nor phone nor a SAML identity, nor a password.

How does this work?

email: If no SAML IdP is present, SCIM user posts must contain an externalId that is an email address. This email address is not added to the newly created user, because it has not been validated. Instead, the flow for changing an email address is triggered in brig: an email is sent to the address containing a validation key, and once the user completes the flow, brig will add the email address to the user. We had to add very little code for this in this PR, it's all an old feature.

When SCIM user gets are processed, in order to reconstruct the externalId from the user spar is retrieving from brig, we introduce a new json object for the sso_id field that looks like this: {'scim_external_id': 'me@example.com'}.

In order to find users that have email addresses pending validation, we introduce a new table in spar's cassandra called scim_external_ids, in analogy to user. We have tried to use brig's internal GET /i/user&email=..., but that also finds pending email addresses, and there are corner cases when changing email addresses and waiting for the new address to be validated and the old to be removed... that made this approach seem infeasible.

password: once the user has validated their email address, they need to trigger the "forgot password" flow -- also old code.

That's the gist of it!

* New brig internal end-points.
* Support for email/password-authenticated scim users.
* Fix: spar's notion of brig's api.
- use internal end-points
- return 'UserAccount' (contains status) instead of 'User'.
- more transparent error handling.
* Fix: allow UserSSOId in brig to carry scim external ids.
* Fix: UserSSOId parsing in spar.
* Fix: store RichInfo in brig if it changes.
* Fix: do not support setting passwords in SCIM
* Fix: update sso_id in brig correctly.
* Fix: do not pull users with email pending validation.
* Refactor: functions for handler-, email-based scim user lookup.
* Give externalIds that are emails their own lookup table in spar.
* Fix: check if email address is available (even without idp).
* Fix: scim-delete if there is no saml idp.
* Refactor: reduce integration test setup time.
* Fix: do not use email address as saml subject if no uref is found.
* Add at least a few lines of docs.
* Assert that deleteUser returns 204


Co-authored-by: Arian van Putten <arian.vanputten@gmail.com>