Fix email validation for saml credentials.
The `doValidate True` case was left-over noise from the old flow,
before this PR.  We only want to send email validation emails if there
is a saml identity involved; if scim invites user to auth with
password, we want to send follow the old invitation flow (no email
needed here in that case).

Technically, the behavior doesn't change with this commit (because
`doValidate True` will always encounter `PendingInvitation` users,
which are not found by `getBrigUserTeam` here, so nothing happens.
but the code is a lot less confusing and misleading.