Wild guess: do templates keep being loaded?
I don't have a theory why the changes here would help, but I think
it's suspicious that we're loading lists from the file system in IO,
and then only forcing those lists a lot later.  And more suspicious,
that a function that should only be called at boot time is
accumulating memory linearly.  But maybe I'm reading the graphs wrong?