brig: Simplify SRV record fetching
The DNS library that we are using already does caching and honors TTLs
published in DNS. The important part to making this work is that all
calls to `resolve` should use the same `Resolver`

In the `dns-util` library that we used before, we would create a new
`Resolver` object for each request, side-stepping the caching
implementation completely.

By reusing the `Resolver` across threads, calls to `resolveSRV` will
only resolve when a cached entry is expired. The default maximum TTL is
300 seconds.

Now there is also no need for running a worker thread in the background
fetching the record periodically.

I have more faith in this library implementing caching correctly than I
have faith in our implementation and it's less code to maintain.

libs/dns-util: remove DNS/Effect.hs

The function is buggy as it allocates a new resolver for every call,
whilst you should reuse the resolver across calls.

Co-Authored-By: Florian Klink <flokli@flokli.de>