[charts] charts/sftd: introduce (wireapp/wire-server-deploy#382)
* charts/sftd: introduce

Adds a new helm chart to deploy sftd as a single replica.

This is suitable for small installations with <50 users.
In the future we will support running SFTD with multiple replicas to
support scale-out of conference calling infrastructure.

The chart runs sftd in the hostNetwork:true such that the SFT
can allocate WebRTC mediastreams on its public IP so people can connect
and call.

This is not ideal, but it's the best workaround I could come up with.

We could solve this by getting
https://github.com/kubernetes/kubernetes/issues/23864 fixed but I don't
have a lot of hope as it has been open for quite a few years.

Other way to fix this is to not use IPv4 but Ipv6 in the cluster. Then
every pod has a proper public ip address and non of this is an issue at
all to begin with. But given IPv6 adoption isn't particularly high,
and because I don't think our TURN servers could bridge the IPv4-IPv6
gap I do not think it's a realistic option for now.

* charts/sftd: Detect the public IP of the (hostNetwork) pod

This information isn't static, as the pod might move around between
nodes. So we need to dynamically discover this.

This won't work in environments behind 1:1 NAT

From discussion with the Calling team:

SO, the  binding address will be the address of the default network
interface, i.e. whatever interface is used for the default gateway, and
has UP status.  essentially it uses either the route list, if the OS
supports it, and chooses the interface that is used for the default
gateway route, OR if this is not available, then what getifaddrs returns
as the first interface that has the status "UP"

* ansible/kubernetes: Support annotating nodes

This is useful to provide information from ansible to be available in
kubernetes.

For example if you want to annotate the node with its external IP
on-prem:

```yaml
node_annotations:
  wire.com/access-ip: {{ access_ip }}
```

* charts/sftd:  Discover external IP of node we're running on

This is useful when sft is run behind 1:1 NAT

* charts/brig:  Point to SFT server