unit test and fix for null values in rendered JSON in UserProfile (#1292)
* unit test and fix for null values in rendered JSON in UserProfile

While omitting a json field, or setting the field to null parses the
same for Haskell, that is not the case for other programming languages,
which can easily trip up over a change in behaviour.

The problem was accidentally introduced in #1281, resulting in null
values instead of omitting fields for UserProfile json, e.g.

{
   "deleted" : null,
   "picture" : [],
   "email" : null,
   "accent_id" : 1,
   "team" : null,
   "handle" : "test1",
   "expires_at" : null,
   "locale" : "en-US",
   "qualified_id" : {
      "id" : "d93f2555-e4b7-4051-8d84-2bcf59ea6066",
      "domain" : "staging.zinfra.io"
   },
   "service" : null,
   "name" : "test",
   "id" : "d93f2555-e4b7-4051-8d84-2bcf59ea6066",
   "assets" : [
      {
         "type" : "image",
         "size" : "preview",
         "key" : "3-1-43978d07-040c-4f49-be7e-06e2560701fd"
      },
      {
         "type" : "image",
         "key" : "3-1-f3039a69-a0b8-45a5-bc6b-3da087b680c7",
         "size" : "complete"
      }
   ]
}

This caused a problem as described in https://wearezeta.atlassian.net/browse/SQCORE-351

This PR reverts the offending change, and adds one manual unit test for
this particular data type. We may wish to add more tests for other data
types as well in future PRs to catch some backwards compatibility changes.