Wire.API.UserMap: Generate nicer example using Arbitrary
By default swagger-ui would generate "additionalProp1", "additionalProp2" .. for
keys in each `Map` type. To get around this we provide examples explicitly. This
results in weirdly named domains like '8v16tdys.s' being generated as well as
empty lists but it seems to be better than "additionalProp1". This causes some
of the lists in the examples to be empty, which is not very helpful as an
example, perhaps there is a way to tweak this.