Brig: Return correct status phrase and body on error
Previously all the servant endpoints were wrongly fabricating the phrase for
status and only putting the message in the body. So, some 404's would return
something like:

```
HTTP/1.1 404 handle-not-found
```

This is allowed in HTTP, but still is a breaking change to our API. So, I think
it should be reverted.

Apart from this we also broke our API by not writing the whole error object in
the body, fortunately nothing broke, but I am pretty sure something would have
broken if we touched the endpoints where clients try to parse the error labels
and messages.