Improve Ctrl-C detection of ormolu script
It seems that ormolu returns exit code 0 on SIGTERM, and this causes the ormolu
script to not react to Ctrl-C, if the ormolu process happens to receive it (as
opposed to the rest of the shell script).

To fix this, run ormolu in background, and wait on the corresponding process.
This makes the wait call handle the interrupt, so that we can distinguish
between normal failure (exit code 100), success (exit code 0) and abnormal
failure.