Internal end-point for ejpd request processing. (#1484)
* Internal end-point for ejpd request processing.

* Fix integration tests on concourse.

* Catch all connections (not just accepted ones).

* Add missing roundtrip test.

In my (poor) defense, it was tested, but via a containing type in
another package.

* Extend swagger.

* Fix integration test.

we have to config files in brig integration tests: one for integration
tests, and one for the brig service itself.  locally, we use
brig.integration.yaml for both and make sure that it contains the
fields for both parsers.  this caused the brig-integration executable
to find brig locally.

further, there is an end-point for brig in the brig service config in
our helm charts.  this caused the brig-integration executable to look
for brig under localhost in the CI, which failed.

looking for the brig end-point in the integration config, not the
service config, should fix things.

* More swagger.