Implement NoConsent semantics on TeamMember.
(I considered changing behavior of addTeamMember as well, but I'm not
sure that'd be an improvement.  I'm trying to avoid migrations, so I
read the whitelist from the config file on every lookup.  This would
still be necessary if *some* of the team members in whitelisted teams
have their record in cassandra straightened out, so I feel it would
just add confusion and not help much.)