Add required fields to Swagger for SchemaP (#1536)
* Add required fields to Swagger schemas

There are two parts to this:

  1) SchemaDoc is now an instance of a NearSemiRing type class, to
  support the existence of two different schema concatenation operations.
  The second operation `add` is used by the Alternative instance, and
  the main one `(<>)` continues to be used by the Applicative one. The
  Swagger implementation of `add` works the same as `(<>)`, but it
  intersects required fields instead of concatenating them.

  2) Required fields are added to the Swagger doc by the `field`
  combinator. Single fields are always required, whereas trivial fields
  obtained with say `pure` are not. Since the `Alternative` instance
  intersects required fields, we automatically get the correct behaviour
  that fields with a default value are not required, as well as fields
  that use the `optional` combinator from `Control.Applicative`.

Unfortunately, the NearSemiRing instance for Swagger is not exactly
law-abiding, but this comes down to Swagger not really being designed to
support schema alternatives. In special situations, one might need to
tweak the required field of the resulting schema, but the current
solution should cover most cases.

* Use Monoid instance for element SchemaP

The Alternative instance used to work as well (and have the same
implementation), but after the introduction of the NearSemiRing
constraint for doc, the Monoid instance is the only one that can be
used in this case.