Merge Empty200, Empty404 and EmptyResult
There is a slight complication here due to the fact that Empty200 and
Empty404 are used in a `UVerb`, and `UVerb` does not really support the
no content type case.

The solution in this patch is to define an overlapping instance for
`HasServer UVerb` for an empty list of content types.

Also, the combinators in `ErrorDescription` (taken from
`servant-swagger`) have been moved to a new module and reused for the
definition of the `Swagger` instance for a `UVerb` contaning empty
results.